Refactored ResourceExecutableExtensionFactory to remove dependency to SText
