+++ plugins/org.yakindu.sct.model.resource/src/org/yakindu/sct/model/resource/SCTResourceDescriptionStrategy.java
@@ -8,7 +8,7 @@
  * 	committers of YAKINDU - initial API and implementation
  * 
  */
-package org.yakindu.sct.model.sgraph.resource.provider;
+package org.yakindu.sct.model.resource;
 
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.util.EcoreUtil;
