+++ plugins/org.yakindu.sct.model.resource/src/org/yakindu/sct/model/resource/validation/SCTResourceValidatorImpl.java
@@ -8,16 +8,16 @@
  * 	committers of YAKINDU - initial API and implementation
  * 
  */
-package org.yakindu.sct.model.stext.ui.validation;
+package org.yakindu.sct.model.resource.validation;
 
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.resource.Resource;
-import org.eclipse.gmf.runtime.notation.Diagram;
 import org.eclipse.xtext.util.CancelIndicator;
 import org.eclipse.xtext.util.IAcceptor;
 import org.eclipse.xtext.validation.CheckMode;
 import org.eclipse.xtext.validation.Issue;
 import org.eclipse.xtext.validation.ResourceValidatorImpl;
+import org.yakindu.sct.model.sgraph.Statechart;
 import org.yakindu.sct.model.sgraph.resource.AbstractSCTResource;
 
 /**
@@ -38,7 +38,7 @@ public class SCTResourceValidatorImpl extends ResourceValidatorImpl {
 	protected void validate(Resource resource, final CheckMode mode, final CancelIndicator monitor,
 			IAcceptor<Issue> acceptor) {
 		for (EObject ele : resource.getContents()) {
-			if (ele instanceof Diagram) {
+			if (!(ele instanceof Statechart)) {
 				continue;
 			}
 			if (!monitor.isCanceled()) {
