+++ plugins/org.yakindu.sct.model.sgraph/META-INF/MANIFEST.MF
@@ -11,7 +11,6 @@ Export-Package: org.yakindu.sct.model.sgraph,
  org.yakindu.sct.model.sgraph.impl,
  org.yakindu.sct.model.sgraph.naming,
  org.yakindu.sct.model.sgraph.resource,
- org.yakindu.sct.model.sgraph.resource.provider,
  org.yakindu.sct.model.sgraph.util,
  org.yakindu.sct.model.sgraph.validation
 Require-Bundle: org.eclipse.core.runtime,
