+++ test-plugins/org.yakindu.sct.generator.java.test/META-INF/MANIFEST.MF
@@ -12,6 +12,8 @@ Require-Bundle: org.yakindu.sct.test.models,
  org.yakindu.sct.generator.genmodel,
  org.eclipse.jdt.core,
  org.eclipse.ui,
- org.eclipse.core.resources
+ org.eclipse.core.resources,
+ org.eclipse.xtext.ui.shared,
+ org.yakindu.sct.generator.builder
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-Vendor: statecharts.org
