+++ test-plugins/org.yakindu.sct.model.stext.resource.test/META-INF/MANIFEST.MF
@@ -6,5 +6,6 @@ Bundle-Version: 2.7.1.qualifier
 Fragment-Host: org.yakindu.sct.model.resource
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Require-Bundle: org.junit,
- org.yakindu.sct.test.models
+ org.yakindu.sct.test.models,
+ org.yakindu.sct.model.stext.ui
 Bundle-Vendor: statecharts.org
